refactor: cache -> ttlcache
