--- cache/doc.go
@@ -1,2 +1,0 @@
-// Package cache provides cache implementations.
-package cache
