+++ ttlcache/doc.go
@@ -1,0 +1,2 @@
+// Package ttlcache provides TTL cache implementations.
+package ttlcache
