+++ ttlcache/memory_test.go
@@ -1,0 +1,11 @@
+package ttlcache
+
+import (
+	"testing"
+)
+
+func TestTLLInMemoryCache(t *testing.T) {
+	testCache(t, func() Cache[int] {
+		return NewInMemory[int]()
+	})
+}
