+++ ttlcache/otel_test.go
@@ -1,0 +1,11 @@
+package ttlcache
+
+import (
+	"testing"
+)
+
+func TestTTLOtel(t *testing.T) {
+	testCache(t, func() Cache[int] {
+		return WithOtel(NewInMemory[int](), "test")
+	})
+}
